fix(public): lightbox + archive bugs found on beta
Three issues surfaced testing the photo lightbox on beta (Epic 7 frontend vs the
prod backend, which lacks Epic 7 endpoints):

1. Past events forced login. SessionQnaThread (Story 7.5) fetches the per-session
   Q&A on public archive pages; on the prod backend that endpoint 401s, and the
   global axios interceptor redirected to /login — bouncing anonymous visitors off
   public archive pages. The Q&A GET is a public, optional read: mark it
   `skipAuthRedirect` so a 401 no longer force-logs-out the page (the query already
   renders empty on error). Hardens prod too — a public archive read must never
   redirect to login.

2. Homepage photos weren't clickable. The homepage strip is TestimonialSection's
   "recent event photos" row, a different marquee than the event page's. Extracted
   the clickable-marquee + lazy-lightbox into a shared ClickablePhotoMarquee and
   used it in both EventPhotosMarquee and the homepage row.

3. Lightbox broken images. The CDN resize Lambda 503s for high-megapixel sources
   above ~1200px (verified on BATbern58: w=1600 → 503, w=512 thumb → 200), so the
   w=1600 slide src showed broken images. The lightbox now serves the raw origin
   (always 200, best zoom quality); marquee thumbnails stay CDN-resized (w=512).
   TODO(infra): give the image-resize Lambda headroom for large sources, then
   switch the lightbox back to a resized width.

tsc + eslint + prettier clean; marquee + qna + SessionCards tests green. [no-doc]
